requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
psycopg2==2.9.3
python-dotenv==0.20.0
PyJWT==2.4.0
bcrypt==3.2.2